Move smog proptotypes to heads
